+++ spec/dummy/src/config/configuration.peakflow.sqlite.js
@@ -52,6 +52,7 @@ export default new Configuration({
       }
     }
   },
+  cookieSecret: "dummy-cookie-secret",
   directory: dummyDirectory(),
   environmentHandler: new NodeEnvironmentHandler(),
   initializeModels: async ({configuration}) => {